djangoapp/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.7.2
boto3==1.35.42
botocore==1.35.42
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
h11==0.14.0
hashids==1.3.1
httpcore==1.0.5
httpx==0.27.2
idna==3.7
jiter==0.5.0
jmespath==1.0.1
mercadopago==2.2.3
openai==1.48.0
packaging==24.1
pillow==10.2.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
replicate==0.34.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
